requirements.txt
asgiref==3.2.10
Django~=2.2.12
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-environ==0.4.5
django-js-asset==1.2.2
django-taggit==1.3.0
Pillow~=7.0.0
pytz==2020.1
sqlparse==0.3.1
